#13 Refactor notifications
